Revert "Page ressource"
