+++ src/app/Admin/base-layout-admin/base-layout-admin.page.html
@@ -1,3 +1,4 @@
+<link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css" rel="stylesheet">
 <div class="flex h-screen bg-gray-100">
   <div [ngClass]="{'hidden': !isSidebarOpen}"
     class="w-64 bg-gradient-to-b from-gray-800 to-gray-900 text-white flex flex-col shadow-xl transition-all duration-300">
