--- src/app/Admin/components/add-resource-modal/add-resource-modal.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { AddResourceModalComponent } from './add-resource-modal.component';
-
-describe('AddResourceModalComponent', () => {
-  let component: AddResourceModalComponent;
-  let fixture: ComponentFixture<AddResourceModalComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ AddResourceModalComponent ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(AddResourceModalComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
