--- src/app/Admin/components/add-resource-modal/add-resource-modal.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { IonicModule } from '@ionic/angular';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
-import { AddResourceModalComponent } from './add-resource-modal.component';
-
-@NgModule({
-  declarations: [
-    AddResourceModalComponent
-  ],
-  imports: [
-    CommonModule,
-    IonicModule,
-    FormsModule,
-    ReactiveFormsModule
-  ],
-  exports: [
-    AddResourceModalComponent
-  ]
-})
-export class AddResourceModalModule {}
