--- src/app/Admin/components/resource-card/resource-card.component.ts
@@ -1,12 +1,0 @@
-import { Component, Input } from '@angular/core';
-import { Resource } from '../../models/resource.model';
-
-@Component({
-  selector: 'app-resource-card',
-  templateUrl: './resource-card.component.html',
-  styleUrls: ['./resource-card.component.scss'],
-  standalone: false,
-})
-export class ResourceCardComponent {
-  @Input() resource!: Resource;
-}
