--- src/app/Admin/components/resource-card/resource-card.module.ts
@@ -1,18 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { IonicModule } from '@ionic/angular';
-import { ResourceCardComponent } from './resource-card.component';
-
-@NgModule({
-  declarations: [
-    ResourceCardComponent
-  ],
-  imports: [
-    CommonModule,
-    IonicModule
-  ],
-  exports: [
-    ResourceCardComponent
-  ]
-})
-export class ResourceCardModule {}
