--- src/app/Admin/models/resource.model.ts
@@ -1,7 +1,0 @@
-export interface Resource {
-  id: string;
-  title: string;
-  description: string;
-  type: 'Zikrs' | 'Livres' | 'Conférences' | 'Tafsirs';
-  dateAdded: string;
-}
