+++ src/app/Admin/pages/dahira/dahira.page.html
@@ -37,7 +37,7 @@
   <!-- Grid Layout -->
   <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
     <!-- Card 1 -->
-    <div (click)="startConfetti()"
+    <div (click)="triggerConfetti()"
       class="bg-white rounded-lg overflow-hidden shadow-sm border border-gray-200 hover:shadow-md transition-shadow duration-300">
       <div class="p-5 flex justify-between items-start" >
         <div>
