--- src/app/Admin/pages/demandes/model/demandes.interface.ts
@@ -1,10 +1,0 @@
-export interface Demande {
-    id: number;
-    demandeur: string;
-    role: string;
-    type: string;
-    sujet: string;
-    date: string;
-    statut: 'En attente' | 'Approuvée' | 'Rejetée';
-    motifRejet?: string;
-}
