--- src/app/Admin/pages/ressources/ressources-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { RessourcesPage } from './ressources.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: RessourcesPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class RessourcesPageRoutingModule {}
