--- src/app/Admin/pages/ressources/ressources.page.scss
@@ -1,11 +1,0 @@
-.custom-searchbar {
-  --background: #f3f4f6;
-  --color: #374151;
-  --placeholder-color: #9ca3af;
-  --icon-color: #9ca3af;
-  --border-radius: 0.375rem;
-}
-
-ion-content {
-  --background: #f9fafb;
-}
