+++ src/index.html
@@ -14,8 +14,7 @@
   <meta name="msapplication-tap-highlight" content="no" />
 
   <link rel="icon" type="image/png" href="assets/icon/favicon.png" />
-  <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css" rel="stylesheet">
-  <!-- <link href="https://cdn.jsdelivr.net/npm/preline@1.x/dist/preline.min.css" rel="stylesheet"> -->
+  <link href="https://cdn.jsdelivr.net/npm/preline@1.x/dist/preline.min.css" rel="stylesheet">
 
   <!-- add to homescreen for ios -->
   <meta name="mobile-web-app-capable" content="yes" />
@@ -24,8 +23,9 @@
 
 <body>
   <app-root></app-root>
+  <!-- CSS -->
   <!-- JS -->
-  <!-- <script src="https://cdn.jsdelivr.net/npm/preline@1.x/dist/preline.min.js"></script> -->
+  <script src="https://cdn.jsdelivr.net/npm/preline@1.x/dist/preline.min.js"></script>
 </body>
 
 </html>
